refactor inquiry to service
