+++ main/views/Total_Email_of_Language/services/Total_Email_of_Language.py
@@ -1,6 +1,7 @@
 from django.http import JsonResponse
 from main.models import UploadedFile
-from main.views.inquiry import get_total_languages_summary, cal_inquiry 
+from main.views.inquiry import get_total_languages_summary
+from main.services.inquiry_service import InquiryService
 from main.views.feedback_package import cal_FeedbackAndPackage
 from main.views.appointment import find_appointment_from_csv_folder
 from main.views.percentage.cal_percentage import find_percentage, cal_percent
@@ -13,7 +14,7 @@ def cal_TotalMonth(date, Web_Commerce):
 
         # print(dateset1, dateset2)
         # total_inquiry = get_total_languages_summary(date)
-        total_inquiry, chart = cal_inquiry(dateset1, dateset2)
+        total_inquiry, chart = InquiryService.cal_inquiry(dateset1, dateset2)
         total_feedback_package = cal_FeedbackAndPackage(date)
         total_appointment = find_appointment_from_csv_folder((dateset1, dateset2))
